Chinese New Year

+++ Content/Buffs/Werewolf.cs
@@ -15,8 +15,7 @@ namespace Consolaria.Content.Buffs
 
 		public override void Update(Player player, ref int buffIndex){ 
 			player.buffTime[buffIndex] = 18000;
-			sbyte type = (sbyte)ModContent.ProjectileType<Projectiles.Friendly.Pets.Werewolf>();
-			//var entitySource = player.GetSource_Buff(buffIndex);
+			ushort type = (ushort)ModContent.ProjectileType<Projectiles.Friendly.Pets.Werewolf>();
 			if (player.whoAmI == Main.myPlayer && player.ownedProjectileCounts[type] <= 0)
 				Projectile.NewProjectile(player.GetSource_Buff(buffIndex), player.Center, Vector2.Zero, type, 0, 0f, player.whoAmI);
 		}
